add lib/processing/vector.rb
